+++ src/ts/wysiwyg/index.ts
@@ -224,7 +224,8 @@ class WYSIWYG {
             }
 
             if ((startSpace && !blockElement.querySelector(".language-mindmap"))
-                || endSpace || isHrMD(blockElement.innerHTML) || isHeadingMD(blockElement.innerHTML)) {
+                || endSpace || isHrMD(blockElement.innerHTML) ||
+                isHeadingMD(blockElement.innerHTML, vditor.options.preview.markdown.setext)) {
                 return;
             }