+++ CHANGELOG.md
@@ -68,6 +68,8 @@
 
 ### v3.3.3 / 2020-06-xx
 
+* [514](https://github.com/Vanessa219/vditor/issues/514) 只读没有禁用场景菜单 `修复缺陷`
+* [501](https://github.com/Vanessa219/vditor/issues/501) 反斜杠 \ 转义问题 `改进功能`
 * [508](https://github.com/Vanessa219/vditor/issues/508) 软键盘 bug `修复缺陷`
 * [504](https://github.com/Vanessa219/vditor/issues/504) 解决编辑器中 esc 推出 hint 和 options.esc 冲突 `修复缺陷`
 * [500](https://github.com/Vanessa219/vditor/issues/500) wysiwyg 模式下打字数学公式会闪 `改进功能`